Fix bug with error positioning
There was a bug where errors were not aligned with each test case. Also, the error message with the Happy New Years one didn't work, because for one, for-in loops over the INDICES of an array, not the items (for-of loops over items), and even with for-of, it won't be overwritten by setting the variable. I fixed this.